pyproject.toml
[project]
name = "Hidden-State-Genomics"
dynamic = ["version"]
description = "A project to identify interpretable features in language model embeddings of nucleotide sequences, and develop explainable AI-systems for variant effect prediction."
authors = [
    { name = "Eliot Kmiec", email = "ek990@georgetown.edu" },
    { name = "Matthew McCoy", email = "mdm299@georgetown.edu" }
]
license = { text = "CC0-1.0" }
#I was having issues installing some dependencies in a more recent version of python. -sob
requires-python = "==3.12.8"

dependencies = [
    "torchvision==0.20.1",
    "torch==2.5.1",
    "nnsight==0.3.7",
    "transformers==4.47.1",
    "biopython==1.84",
    "hgvs==1.5.4",
    "biocommons-seqrepo==0.6.9",
    "pandas==2.2.3",
    "numpy==2.0.0",
    "tqdm==4.67.1",
    "seaborn==0.13.2",
    "plotly==5.24.1",
    "scikit-learn==1.6.0",
    "umap-learn==0.5.7",
    "streamlit==1.41.1",
    "google-cloud-bigquery==3.27.0",
    "typed-argument-parser==1.10.1",
    "python-dotenv==1.0.1",
    "pyspark==3.5.4",
    "gcsfs==2024.12.0",
    "tensorboard==2.19.0",
    "jupyter",
    "datashader",
    "bokeh",
    "holoviews",
]

[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["hsg"]